+++ lib/src/readers/id3v2.dart
@@ -1,12 +1,12 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:dart_tags/src/model/consts.dart' as consts;
-import 'package:dart_tags/src/model/attached_picture.dart';
+import 'package:dart_tags/src/frames/frame.dart';
 import 'package:dart_tags/src/readers/reader.dart';
 
 class ID3V2Reader extends Reader {
   // [ISO-8859-1]. Terminated with $00.
+  // ignore: unused_field
   static const _latin1 = 0x00;
 
   // [UTF-16] encoded Unicode [UNICODE] with BOM. All strings in the same frame SHALL have the same byteorder. Terminated with $00 00. (use in future)
@@ -18,6 +18,7 @@ class ID3V2Reader extends Reader {
   static const _utf16be = 0x02;
 
   // [UTF-8] encoded Unicode [UNICODE]. Terminated with $00.
+  // ignore: unused_field
   static const _utf8 = 0x03;
 
   int get _headerLength => version_o2 > 2 ? 10 : 6;
@@ -43,6 +44,8 @@ class ID3V2Reader extends Reader {
     version_o2 = sBytes[3];
     version_o3 = sBytes[4];
 
+    final ff = FrameFactory('ID3', '2.4.0');
+
     final flags = sBytes[5];
 
     // ignore: unused_local_variable
@@ -59,55 +62,18 @@ class ID3V2Reader extends Reader {
     var end = true;
 
     while (end) {
-      final encoding = _getEncoding(sBytes[offset + _headerLength]);
-      final tag = encoding.decode(sBytes.sublist(offset, offset + 4));
-
       final len = _sizeOf(sBytes.sublist(offset + 4, offset + 8));
+      final fr = sBytes.sublist(offset);
 
-      if (len > 0 && _isValidTag(tag)) {
-        switch (tag) {
-          case 'APIC':
-            tags[tag] = _parsePicture(
-                sBytes.sublist(
-                    offset + _headerLength + 1, offset + _headerLength + len),
-                encoding);
-            break;
-          case 'TXXX':
-            final frame = sBytes.sublist(
-                offset + _headerLength + 1, offset + _headerLength + len);
-            final splitIndex = frame.indexOf(0);
-            tags[splitIndex == 0
-                    ? tag
-                    : encoding.decode(frame.sublist(0, splitIndex))] =
-                encoding.decode(frame.sublist(splitIndex + 1));
-            break;
-          case 'WXXX':
-            final frame = sBytes.sublist(
-                offset + _headerLength + 1, offset + _headerLength + len);
-            final splitIndex = frame.indexOf(0);
-            tags[splitIndex == 0
-                    ? tag
-                    : encoding.decode(frame.sublist(0, splitIndex))] =
-                latin1.decode(frame.sublist(splitIndex + 1));
-            break;
-          default:
-            tags[_getTag(tag)] = encoding.decode(_clearFrameData(sBytes.sublist(
-                offset + _headerLength + 1, offset + _headerLength + len)));
-        }
-      }
+      final m = ff.getFrame(fr).decode(sBytes.sublist(offset));
+
+      tags[m?.key] = m?.value;
 
       offset = offset + _headerLength + len;
       end = offset < size;
     }
 
     return tags;
-  }
-
-  List<int> _clearFrameData(List<int> bytes) {
-    if (bytes.length > 3 && bytes[0] == 0xFF && bytes[1] == 0xFE) {
-      bytes = bytes.sublist(2);
-    }
-    return bytes.where((i) => i != 0).toList();
   }
 
   int _sizeOf(List<int> block) {
@@ -119,81 +85,5 @@ class ID3V2Reader extends Reader {
     len += block[3];
 
     return len;
-  }
-
-  Encoding _getEncoding(int type) {
-    switch (type) {
-      case _latin1:
-        return latin1;
-      case _utf8:
-        return Utf8Codec(allowMalformed: true);
-      default:
-        return UTF16();
-    }
-  }
-
-  String _getTag(String tag) {
-    return consts.frameHeaderShortcutsID3V2_3.containsKey(tag)
-        ? consts.frameHeaderShortcutsID3V2_3[tag]
-        : tag;
-  }
-
-  bool _isValidTag(String tag) {
-    return consts.framesHeaders.containsKey(tag);
-  }
-
-  AttachedPicture _parsePicture(List<int> sublist, Encoding enc) {
-    final iterator = sublist.iterator;
-    var buff = <int>[];
-
-    final attachedPicture = AttachedPicture();
-
-    var cont = 0;
-
-    while (iterator.moveNext() && cont < 4) {
-      final crnt = iterator.current;
-      if (crnt == 0x00 && cont < 3) {
-        if (cont == 1 && buff.isNotEmpty) {
-          attachedPicture.imageTypeCode = buff[0];
-          cont++;
-          attachedPicture.description = enc.decode(buff.sublist(1));
-        } else {
-          attachedPicture.mime = enc.decode(buff);
-        }
-        buff = [];
-        cont++;
-        continue;
-      }
-      buff.add(crnt);
-    }
-
-    attachedPicture.imageData = buff;
-
-    return attachedPicture;
-  }
-}
-
-class UTF16 extends Encoding {
-  @override
-  Converter<List<int>, String> get decoder => _UTF16Decoder();
-
-  @override
-  Converter<String, List<int>> get encoder => _UTF16Enoder();
-
-  @override
-  String get name => 'utf16';
-}
-
-class _UTF16Decoder extends Converter<List<int>, String> {
-  @override
-  String convert(List<int> input) {
-    return String.fromCharCodes(input);
-  }
-}
-
-class _UTF16Enoder extends Converter<String, List<int>> {
-  @override
-  List<int> convert(String input) {
-    return input.runes.toList();
   }
 }